created a function which only handles calculations of exp for each value of a member returns those values as an object to be handled accordingly
