+++ markdown/bitburner.gangmemberexpgain.agiexp.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [bitburner](./bitburner.md) &gt; [GangMemberExpGain](./bitburner.gangmemberexpgain.md) &gt; [agiEXP](./bitburner.gangmemberexpgain.agiexp.md)
+
+## GangMemberExpGain.agiEXP property
+
+Agi EXP gain for a member after a given task
+
+**Signature:**
+
+```typescript
+agiEXP: number;
+```
