+++ markdown/bitburner.gangmemberexpgain.chaexp.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [bitburner](./bitburner.md) &gt; [GangMemberExpGain](./bitburner.gangmemberexpgain.md) &gt; [chaEXP](./bitburner.gangmemberexpgain.chaexp.md)
+
+## GangMemberExpGain.chaEXP property
+
+Cha EXP gain for a member after a given task
+
+**Signature:**
+
+```typescript
+chaEXP: number;
+```
