+++ markdown/bitburner.gangmemberexpgain.defexp.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [bitburner](./bitburner.md) &gt; [GangMemberExpGain](./bitburner.gangmemberexpgain.md) &gt; [defEXP](./bitburner.gangmemberexpgain.defexp.md)
+
+## GangMemberExpGain.defEXP property
+
+Def EXP gain for a member after a given task
+
+**Signature:**
+
+```typescript
+defEXP: number;
+```
