+++ markdown/bitburner.gangmemberexpgain.dexexp.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [bitburner](./bitburner.md) &gt; [GangMemberExpGain](./bitburner.gangmemberexpgain.md) &gt; [dexEXP](./bitburner.gangmemberexpgain.dexexp.md)
+
+## GangMemberExpGain.dexEXP property
+
+Dex EXP gain for a member after a given task
+
+**Signature:**
+
+```typescript
+dexEXP: number;
+```
