+++ markdown/bitburner.gangmemberexpgain.hackexp.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [bitburner](./bitburner.md) &gt; [GangMemberExpGain](./bitburner.gangmemberexpgain.md) &gt; [hackEXP](./bitburner.gangmemberexpgain.hackexp.md)
+
+## GangMemberExpGain.hackEXP property
+
+Hack EXP gain for a member after a given task
+
+**Signature:**
+
+```typescript
+hackEXP: number;
+```
