+++ markdown/bitburner.gangmemberexpgain.md
@@ -1,0 +1,24 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [bitburner](./bitburner.md) &gt; [GangMemberExpGain](./bitburner.gangmemberexpgain.md)
+
+## GangMemberExpGain interface
+
+
+**Signature:**
+
+```typescript
+interface GangMemberExpGain 
+```
+
+## Properties
+
+|  Property | Modifiers | Type | Description |
+|  --- | --- | --- | --- |
+|  [agiEXP](./bitburner.gangmemberexpgain.agiexp.md) |  | number | Agi EXP gain for a member after a given task |
+|  [chaEXP](./bitburner.gangmemberexpgain.chaexp.md) |  | number | Cha EXP gain for a member after a given task |
+|  [defEXP](./bitburner.gangmemberexpgain.defexp.md) |  | number | Def EXP gain for a member after a given task |
+|  [dexEXP](./bitburner.gangmemberexpgain.dexexp.md) |  | number | Dex EXP gain for a member after a given task |
+|  [hackEXP](./bitburner.gangmemberexpgain.hackexp.md) |  | number | Hack EXP gain for a member after a given task |
+|  [strEXP](./bitburner.gangmemberexpgain.strexp.md) |  | number | Str EXP gain for a member after a given task |
+
