+++ markdown/bitburner.gangmemberexpgain.strexp.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [bitburner](./bitburner.md) &gt; [GangMemberExpGain](./bitburner.gangmemberexpgain.md) &gt; [strEXP](./bitburner.gangmemberexpgain.strexp.md)
+
+## GangMemberExpGain.strEXP property
+
+Str EXP gain for a member after a given task
+
+**Signature:**
+
+```typescript
+strEXP: number;
+```
